move test helpers into util
