+++ internal/app/cf-terraforming/cmd/util.go
@@ -1,5 +1,15 @@
 package cmd
 
+import (
+	"bytes"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/spf13/cobra"
+)
+
 func contains(slice []string, item string) bool {
 	set := make(map[string]struct{}, len(slice))
 	for _, s := range slice {
@@ -8,4 +18,52 @@ func contains(slice []string, item string) bool {
 
 	_, ok := set[item]
 	return ok
+}
+
+func executeCommandC(root *cobra.Command, args ...string) (c *cobra.Command, output string, err error) {
+	buf := new(bytes.Buffer)
+	root.SetOut(buf)
+	root.SetErr(buf)
+	root.SetArgs(args)
+
+	c, err = root.ExecuteC()
+
+	return c, buf.String(), err
+}
+
+// testDataFile slurps a local test case into memory and returns it while
+// encapsulating the logic for finding it.
+func testDataFile(path string) string {
+	if strings.HasPrefix(path, "/") {
+		path = trimLeftChar(path)
+	}
+
+	dirname, err := os.Getwd()
+	if err != nil {
+		panic(err)
+	}
+
+	dir, err := os.Open(filepath.Join(dirname, "../../../../testdata"))
+	if err != nil {
+		panic(err)
+	}
+
+	fullpath := dir.Name() + "/" + path
+	if _, err := os.Stat(fullpath); os.IsNotExist(err) {
+		panic("path does not exist")
+	}
+
+	data, _ := ioutil.ReadFile(fullpath)
+
+	return string(data)
+}
+
+// trimLeftChar removes the first character of a string.
+func trimLeftChar(s string) string {
+	for i := range s {
+		if i > 0 {
+			return s[i:]
+		}
+	}
+	return s[:0]
 }
